Extract external evidence facade routes
